+++ python/GafferCyclesUI/CyclesAttributesUI.py
@@ -69,7 +69,7 @@ def __subdivisionSummary( plug ) :
 def __volumeSummary( plug ) :
 
 	info = []
-	for childName in ( "volumeClipping", "volumeStepSize", "volumeObjectSpace" ) :
+	for childName in ( "volumeClipping", "volumeStepSize", "volumeObjectSpace", "volumePrecision" ) :
 		if plug[childName]["enabled"].getValue() :
 			info.append( IECore.CamelCase.toSpaced( childName ) + ( " On" if plug[childName]["value"].getValue() else " Off" ) )
 
@@ -338,6 +338,27 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"layout:section", "Volume",
+
+		],
+
+		"attributes.volumePrecision" : [
+
+			"description",
+			"""
+			Specifies volume data precision, lower values reduce
+			memory consumption at the cost of detail.
+			""",
+
+			"layout:section", "Volume",
+
+		],
+
+		"attributes.volumePrecision.value" : [
+
+			"preset:Full", "full",
+			"preset:Half", "half",
+
+			"plugValueWidget:type", "GafferUI.PresetsPlugValueWidget",
 
 		],